+++ tests/sample_app_3/cms_apps.py
@@ -8,14 +8,15 @@ from django.utils.translation import ugettext_lazy as _
 from djangocms_apphook_setup.base import AutoCMSAppMixin
 
 
+@apphook_pool.register
 class App3(AutoCMSAppMixin, CMSApp):
     name = _('App3')
-    urls = ['sample_app_3.urls']
+    _urls = ['sample_app_3.urls']
     auto_setup = {
         'enabled': True,
         'home title': 'home title',
         'page title': 'page 3 title',
     }
 
-apphook_pool.register(App3)
+
 App3.setup()