+++ tests/test_setup.py
@@ -4,13 +4,15 @@ from __future__ import absolute_import, print_function, unicode_literals
 import sys
 
 from cms.api import create_page, create_title
-from cms.models import Page
+from cms.models import Page, Site
 from cms.utils import get_language_list
 from django.utils.translation import override
+from djangocms_helper.utils import CMS_30
 
 from .base import BaseTest
 from .sample_app_1.cms_appconfig import AppConfig
 from .sample_app_4.cms_appconfig import App4Config
+from .sample_app_6.cms_appconfig import App6Config
 
 
 class SetupAppBaseTest(BaseTest):
@@ -20,19 +22,29 @@ class SetupAppBaseTest(BaseTest):
 
     def setUp(self):
         super(SetupAppBaseTest, self).setUp()
-        from cms.apphook_pool import apphook_pool
+        self._delete_modules()
 
+    def _delete_modules(self):
+        from cms.apphook_pool import apphook_pool
         delete = [
             'tests.sample_app_1',
             'tests.sample_app_1.cms_app',
+            'tests.sample_app_1.cms_apps',
             'tests.sample_app_2',
             'tests.sample_app_2.cms_app',
+            'tests.sample_app_2.cms_apps',
             'tests.sample_app_3',
             'tests.sample_app_3.cms_app',
+            'tests.sample_app_3.cms_apps',
             'tests.sample_app_4',
             'tests.sample_app_4.cms_app',
+            'tests.sample_app_4.cms_apps',
             'tests.sample_app_5',
             'tests.sample_app_5.cms_app',
+            'tests.sample_app_5.cms_apps',
+            'tests.sample_app_6',
+            'tests.sample_app_6.cms_app',
+            'tests.sample_app_6.cms_apps',
         ]
         for module in delete:
             if module in sys.modules:
@@ -41,30 +53,31 @@ class SetupAppBaseTest(BaseTest):
             self.config.cmsapp = None
         apphook_pool.clear()
 
-    def _setup_from_cmsapp(self):
+    def _setup_from_cmsapp(self, site_id=1, home_final=4, blog_final=2, configs_final=1, configs_init=0, pages_init=0):
 
-        # Tests starts with no page and no config
-        self.assertFalse(Page.objects.exists())
+        # Tests starts with a set of pages / configs
+        self.assertEqual(Page.objects.count(), pages_init)
         if self.config:
-            self.assertFalse(self.config.objects.exists())
-        self.assertEqual(Page.objects.filter(application_urls=self.app_name).count(), 0)
+            self.assertEqual(self.config.objects.count(), configs_init)
+        self.assertEqual(Page.objects.filter(application_urls=self.app_name).count(), configs_init * 2)
 
         # importing cms_app triggers the auto setup
-        __import__(self.module, fromlist=(str('cms_app'),))
+        __import__(self.module, fromlist=(str('cms_apps'),))
 
-        # Home and blog, published and draft
-        self.assertEqual(Page.objects.count(), 4)
+        # Final set of pages / configs
+        self.assertEqual(Page.objects.count(), home_final)
         if self.config:
-            self.assertEqual(self.config.objects.count(), 1)
-        self.assertEqual(Page.objects.filter(application_urls=self.app_name).count(), 2)
+            self.assertEqual(self.config.objects.count(), configs_final)
+        self.assertEqual(Page.objects.filter(application_urls=self.app_name).count(), blog_final)
 
-    def _setup_filled(self):
-
-        # Tests starts with no page and no config
-        self.assertFalse(Page.objects.exists())
+    def _setup_filled(self, site_id=1, home_final=4, blog_final=2, configs_final=1, configs_init=0, pages_init=0):
+        site = Site.objects.get(pk=site_id)
+        set_home = hasattr(Page, 'set_as_homepage')
+        # Tests starts with a set of pages / configs
+        self.assertEqual(Page.objects.count(), pages_init)
         if self.config:
-            self.assertFalse(self.config.objects.exists())
-        self.assertEqual(Page.objects.filter(application_urls=self.app_name).count(), 0)
+            self.assertEqual(self.config.objects.count(), configs_init)
+        self.assertEqual(Page.objects.filter(application_urls=self.app_name).count(), configs_init * 2)
 
         langs = get_language_list()
         home = None
@@ -72,9 +85,11 @@ class SetupAppBaseTest(BaseTest):
             with override(lang):
                 if not home:
                     home = create_page(
-                        'a new home', language=lang,
+                        'a new home', language=lang, site=site,
                         template='page.html', in_navigation=True, published=True
                     )
+                    if set_home:
+                        home.set_as_homepage()
                 else:
                     create_title(
                         language=lang, title='a new home', page=home
@@ -82,17 +97,17 @@ class SetupAppBaseTest(BaseTest):
                     home.publish(lang)
 
         # importing cms_app triggers the auto setup
-        __import__(self.module, fromlist=(str('cms_app'),))
+        __import__(self.module, fromlist=(str('cms_apps'),))
 
-        # Home and blog, published and draft
-        self.assertEqual(Page.objects.count(), 4)
+        # Final set of pages / configs
+        self.assertEqual(Page.objects.count(), home_final)
         if self.config:
-            self.assertEqual(self.config.objects.count(), 1)
-        self.assertEqual(Page.objects.filter(application_urls=self.app_name).count(), 2)
+            self.assertEqual(self.config.objects.count(), configs_final)
+        self.assertEqual(Page.objects.filter(application_urls=self.app_name).count(), blog_final)
 
-        home = Page.objects.get_home()
+        home = Page.objects.get_home(site)
         for lang in langs:
-            self.assertEqual(home.get_title(lang), 'a new home')
+            self.assertEqual(home.get_title(), 'a new home')
 
 
 class SetupApp1Test(SetupAppBaseTest):
@@ -108,7 +123,7 @@ class SetupApp1Test(SetupAppBaseTest):
 
     def test_config_values(self):
         # importing cms_app triggers the auto setup
-        __import__(self.module, fromlist=(str('cms_app'),))
+        __import__(self.module, fromlist=(str('cms_apps'),))
         config = self.config.objects.first()
         self.assertEqual(config.app_title, 'app title')
 
@@ -148,12 +163,135 @@ class SetupApp4Test(SetupAppBaseTest):
 
     def test_config_values(self):
         # importing cms_app triggers the auto setup
-        __import__(self.module, fromlist=(str('cms_app'),))
+        __import__(self.module, fromlist=(str('cms_apps'),))
         config = self.config.objects.first()
         self.assertEqual(set(config.get_available_languages()), set(('en', 'it', 'fr')))
         self.assertTrue(config.random_option)
         config.set_current_language('en')
         self.assertEqual(config.object_name, 'name')
+
+
+class SetupApp4Site2Test(SetupAppBaseTest):
+    config = App4Config
+    module = 'tests.sample_app_4'
+    app_name = 'App4'
+
+    def setUp(self):
+        super(SetupApp4Site2Test, self).setUp()
+        Site.objects.create(name='domain2', domain='www.example2.com', id=2)
+
+    def test_setup_from_cmsapp(self):
+        with self.settings(SITE_ID=1):
+            self._setup_from_cmsapp(site_id=1, home_final=4, blog_final=2)
+        self._delete_modules()
+        if CMS_30:
+            self.skipTest('Multisite not supported on djago CMS 3.0')
+        else:
+            with self.settings(SITE_ID=2):
+                self._setup_from_cmsapp(
+                    site_id=2, home_final=8, blog_final=4, configs_final=1, configs_init=1, pages_init=4
+                )
+            self.assertEqual(Page.objects.count(), 8)
+            self.assertEqual(Page.objects.on_site(1).count(), 4)
+            self.assertEqual(Page.objects.on_site(2).count(), 4)
+            self.assertEqual(self.config.objects.count(), 1)
+            self.assertEqual(Page.objects.filter(application_urls=self.app_name).count(), 4)
+            self.assertEqual(Page.objects.on_site(1).filter(application_urls=self.app_name).count(), 2)
+            self.assertEqual(Page.objects.on_site(2).filter(application_urls=self.app_name).count(), 2)
+
+    def test_setup_filled(self):
+        with self.settings(SITE_ID=1):
+            self._setup_filled(site_id=1, home_final=4, blog_final=2)
+        self._delete_modules()
+        if CMS_30:
+            self.skipTest('Multisite not supported on djago CMS 3.0')
+        else:
+            with self.settings(SITE_ID=2):
+                self._setup_filled(
+                    site_id=2, home_final=8, blog_final=4, configs_final=1, configs_init=1, pages_init=4
+                )
+            self.assertEqual(Page.objects.count(), 8)
+            self.assertEqual(Page.objects.on_site(1).count(), 4)
+            self.assertEqual(Page.objects.on_site(2).count(), 4)
+            self.assertEqual(self.config.objects.count(), 1)
+            self.assertEqual(Page.objects.filter(application_urls=self.app_name).count(), 4)
+            self.assertEqual(Page.objects.on_site(1).filter(application_urls=self.app_name).count(), 2)
+            self.assertEqual(Page.objects.on_site(2).filter(application_urls=self.app_name).count(), 2)
+
+    def test_config_values(self):
+        # importing cms_app triggers the auto setup
+        with self.settings(SITE_ID=1):
+            __import__(self.module, fromlist=(str('cms_apps'),))
+            config = self.config.objects.first()
+            self.assertEqual(set(config.get_available_languages()), set(('en', 'it', 'fr')))
+            self.assertTrue(config.random_option)
+            config.set_current_language('en')
+            self.assertEqual(config.object_name, 'name')
+        self._delete_modules()
+        if CMS_30:
+            self.skipTest('Multisite not supported on djago CMS 3.0')
+        else:
+            with self.settings(SITE_ID=2):
+                __import__(self.module, fromlist=(str('cms_apps'),))
+                self.assertEqual(self.config.objects.count(), 1)
+                config = self.config.objects.first()
+                self.assertEqual(set(config.get_available_languages()), set(('en', 'it', 'fr')))
+                self.assertTrue(config.random_option)
+                config.set_current_language('en')
+                self.assertEqual(config.object_name, 'name')
+
+
+class SetupApp6Site2Test(SetupAppBaseTest):
+    config = App6Config
+    module = 'tests.sample_app_6'
+    app_name = 'App6'
+
+    def setUp(self):
+        super(SetupApp6Site2Test, self).setUp()
+        Site.objects.create(name='domain2', domain='www.example2.com', id=2)
+
+    def test_setup_from_cmsapp(self):
+        with self.settings(SITE_ID=1):
+            self._setup_from_cmsapp(site_id=1, home_final=0, blog_final=0, configs_final=0)
+        self._delete_modules()
+        with self.settings(SITE_ID=2):
+            self._setup_from_cmsapp(site_id=2, home_final=4, blog_final=2, configs_final=1)
+        self.assertEqual(Page.objects.count(), 4)
+        self.assertEqual(Page.objects.on_site(1).count(), 0)
+        self.assertEqual(Page.objects.on_site(2).count(), 4)
+        self.assertEqual(self.config.objects.count(), 1)
+        self.assertEqual(Page.objects.filter(application_urls=self.app_name).count(), 2)
+        self.assertEqual(Page.objects.on_site(1).filter(application_urls=self.app_name).count(), 0)
+        self.assertEqual(Page.objects.on_site(2).filter(application_urls=self.app_name).count(), 2)
+
+    def test_setup_filled(self):
+        with self.settings(SITE_ID=1):
+            self._setup_filled(site_id=1, home_final=2, blog_final=0, configs_final=0)
+        self._delete_modules()
+        with self.settings(SITE_ID=2):
+            self._setup_filled(site_id=2, home_final=6, blog_final=2, configs_final=1, pages_init=2)
+        self.assertEqual(Page.objects.count(), 6)
+        self.assertEqual(Page.objects.on_site(1).count(), 2)
+        self.assertEqual(Page.objects.on_site(2).count(), 4)
+        self.assertEqual(self.config.objects.count(), 1)
+        self.assertEqual(Page.objects.filter(application_urls=self.app_name).count(), 2)
+        self.assertEqual(Page.objects.on_site(1).filter(application_urls=self.app_name).count(), 0)
+        self.assertEqual(Page.objects.on_site(2).filter(application_urls=self.app_name).count(), 2)
+
+    def test_config_values(self):
+        # importing cms_app triggers the auto setup
+        with self.settings(SITE_ID=1):
+            __import__(self.module, fromlist=(str('cms_apps'),))
+            self.assertEqual(self.config.objects.count(), 0)
+        self._delete_modules()
+        with self.settings(SITE_ID=2):
+            __import__(self.module, fromlist=(str('cms_apps'),))
+            self.assertEqual(self.config.objects.count(), 1)
+            config = self.config.objects.first()
+            self.assertEqual(set(config.get_available_languages()), set(('en',)))
+            self.assertTrue(config.random_option)
+            config.set_current_language('en')
+            self.assertEqual(config.object_name, 'name')
 
 
 class SetupApp5Test(SetupAppBaseTest):
@@ -166,7 +304,7 @@ class SetupApp5Test(SetupAppBaseTest):
         self.assertEqual(Page.objects.filter(application_urls=self.app_name).count(), 0)
 
         # importing cms_app triggers the auto setup
-        __import__(self.module, fromlist=(str('cms_app'),))
+        __import__(self.module, fromlist=(str('cms_apps'),))
 
         # No setup is performed
         self.assertEqual(Page.objects.count(), 0)